Move hdfs validation tool to integration/validation module
This moves the hdfs validation tool into a new module named
`integration/validation`, like `integration/hms`. This enables the tool
to be imported and used where a dependency on `shell` module is not
possible.

Consequently all the related classes have to be moved together with this
tool. That includes the tools for `runUfsTests`, `validateEnv` and
`runHdfsMountTests`. Each one of them is broken into two classes, one
enclosing the real work logic in `integration/`, the other being the
triggering class that only handles the cmdline logic in `shell/`.

pr-link: Alluxio/alluxio#11578
change-id: cid-f56483417bcbad74ca733208fd17bfc86e155a28