Fix deadlock issue on metadata sync
This deadlock occurs when concurrent clients submit requests to
paths where one is the parent of another. It is most likely to
occur one path is a direct parent of another. There needs to be
less threads in the executor than there are paths to sync.
When the deadlock occurs it is because the child path holds a read
lock on a path the parent may be attempting to sync. This prevents
the parent path from continuing to sync. The child path though may
keep queuing sync jobs which can prevent the child from completing
its sync and thus causing deadlock because the parent can never
finish.

In order to fix this there were two major issues to be solved:

1. Unlock the root path so the ancestry of one sync does not block
another. After internal discussion, we decided that because we
don't hold the lock for any children for the duration of the sync
that it is unnecessary to hold the parent lock for the entire
duration as well.

2. Because the inode sync stream has to lock and unlock the root
path while syncing, the function loadMetadataIfNotExists needed
to be modified so that the path is unlocked before loading the
metadata. The major issue is that checking whether we need to
load metadata on a path is expensive because it requires a lock.
In order to maintain high RPC throughput for GetFileInfo and
ListStatus (the users of loadMetadataIfNotExist) this PR
introduces a two-step scheme for locking on those RPCs. The
general approach is to lock and then check if metadata needs to
be loaded, in which case we unlock and go back to the beginning
of the RPCs main logic with an extra flag set to load metdata.
In the case where the path already exists and does not need to be
loaded, then the RPC continues as normal.

There will be a slight overhead in occasions where we are loading
metadata for paths that don't yet exist, but that is not the
common case, so I am slightly less concerned about the performance
there. There will be no more locks taken there than there would be
in a normal metadata sync, so the runtimes should be comparable
anyway.

pr-link: Alluxio/alluxio#11579
change-id: cid-a3b6372531584eb068b4fd2e1c78503dbb80b225