Add minor RPC optimizations
Based on my testing on a local machine these sites accounted for
anywhere between 1% and 3% of the CPU time when processing the RPC.

- When calling getObjectListing on the ObjectUnderFileSystem, we should
store the value of the breadcrumbs enabled as a boolean rather than
always parsing from the config
- When calling `lockShared()` in the state lock manager, prevent a
String.format call by wrapping with a conditional statement.
- When creating the DefaultBlockDeletionContext, initialize the array
list with a size of 0 to prevent unnecessarily allocating small arrays
over and over again with the default ArrayList constructor, when most
times we won't even add to this ArrayList.

pr-link: Alluxio/alluxio#11720
change-id: cid-efe5df25090ff1b70e50a9c07c1110c8bbebc79f