Simplify UfsStatus handling in loadMetadata
This reverts a change from #11704 while also removing redundant UFS
calls. The `ufs.exists` in `loadMetadata` is no longer necessary because
every call to `loadMetadata` goes through call site place which will
always set the `UfsStatus` in the `LoadMetadataContext`. The attempt to
retrieve the `UfsStatus` from the `UfsStatusCache` will always pull the
status from the internal map, or reach out to the UFS. If the status
couldn't be retrieved on the call to `fetchStatusIfAbsent` then it does
not exist in the UFS.

pr-link: Alluxio/alluxio#11746
change-id: cid-9a9def3d491f9e193aa3e3477403b3886498f4fd