Format only the local master to avoid wiping the whole cluster
This prevents the start command from formatting an entire cluster if the
journal directory doesn't exist. Original behavior was built to ensure
that the journal folder existed in EMBEDDED mode. However, if a user
accidentally removes the directory (e.g. during an upgrade), then we
want to ensure that not all nodes are formatted.

pr-link: Alluxio/alluxio#11916
change-id: cid-15ba0851dabb2890f8cae365808dce55322b8f32