Fix local cache locking
Fix two issues:

(1) Previously MetaStore and Evictor are both used as the state to make
eviction decisions, but they are not put in the same critical section,
and not coordinated to be consistent in tricky cases of failed page
store deletes. This PR makes evictor a part of MetaStore so their
content will always be consistent
(2) Eviction is a 2PC (phase 1 in metastore and phase 2 in pagestore),
this PR fixes the undo logic when this 2PC failed in the middle.

pr-link: Alluxio/alluxio#12013
change-id: cid-183d4c93bf7e46105d82ae949c0375c72154d9b4