Add <parent> declaration to webui pom
License check hasn't been properly working for a while on this. Adding
the parent declaration will fix this. It will also allow us to run `mvn
license:check` from the root of the repo. Before, `license:check` would
fail on the webui because it would attempt to use a different plugin.

pr-link: Alluxio/alluxio#12027
change-id: cid-c56d56fe47996cf4fa3463525c2647b7d7e373d4