Update sync status after processing the entire sync
Marking paths as sync during the sync is inaccurate, because the syncs
for descendants are run in a separate thread pool. Therefore, only mark
the root path as synced after everything is finished. This also reduces
the number of entries in the sync cache, instead of marking each path in
the tree.

Fixes #12106

pr-link: Alluxio/alluxio#12133
change-id: cid-2da1ce3223831a44bb04af9e8d40c025fd7108c9