Implement a local first raft client for improved journal performance
Current the raft client for writing journal entries will always checking
for leadership and send the request through network interface,
introducing significant overhead. This change implements a raft client
which attempts sending messages directly to local server first, and
falls back to the default client if the local server is no longer a
leader.

The client also has built-in retry function for raft client which is
closed due to timeout.

pr-link: Alluxio/alluxio#12174
change-id: cid-df05031cb47f6f931e67c9b4c853b9b426ba145d