Fix backup coordination and error handling
This change fixed a few issues related to backup:

- Make backup interrupt correctly when raft journal needs to reload from
a snapshot. Currently if backup occurs right before a journal reload it
can buffer up some entries and then corrupt journal by resuming after
the reload is completed. This change added an interrupt callback to make
sure backup is correctly canceled and journal is resumed before a reload
happens.

- Block backup when journal is being reloaded. When a raft journal is
being reloaded, the backup should not happen because it will capture
inconsistent state while the master metadata is re-applied from
snapshot.

- Cancel backup sub-tasks properly. Currently when a backup is cancelled
due to error or interruption, the sub-tasks do not explicitly respond to
interrupt signals, leaving them running in background while opening
backup service for more requests. This causes increased memory
consumption when multiple backup tasks are running concurrently.

- Incomplete backups. There is currently no way to tell if a backup file
is incomplete before restoring it to the cluster. This change added a
marker file after the backup is done to indicate the completeness.

pr-link: Alluxio/alluxio#12214
change-id: cid-17276237434626320356c5896b7cade1900ed9f1