Do not configure alluxio catalog for older versions of presto in emr
in emr 5.25, which we refer to in our docs, the presto version is 0.220.
the connector we bundle with the tarball only works with 0.227+. if we
configure the alluxio catalog that uses our bundled connector and the
presto version is incompatible, presto will fail to start.

this change checks for the emr version and ensures that it is >= 5.28,
which is the first version that uses presto 0.227. here we are assuming
that presto will be backwards compatible

i have tested this with both emr 5.25 and 5.28 to ensure the catalog is
configured as described

pr-link: Alluxio/alluxio#12283
change-id: cid-e4013a4c25bf330ca9dd4e1f610e0468ce81940d