Fix backup worker crashing master on interruption
When a backup is interrupted, there is a chance that the resume callback
cannot be canceled and thus happens after the server is already resumed
by the journal reload process. While it makes sense for backup to crash
the master if resume fails to complete and master cannot accept new
journal entries, it is completely fine for master to continue if backup
tries to resume an already resumed journal. This change updates the
journal to not throw if someone attempts to resume after the journal is
already resumed.

pr-link: Alluxio/alluxio#12269
change-id: cid-89f233a77b23acd8655c335a5f317cae1c5d62f0