Add a releaseProfile to generate dependency reduced pom
#12338

This is a workaround for MSHADE-148, which leads to an infinite loop
when building Alluxio.

There are two scenarios.
- Releasing Alluxio
Maven build with `-Prelease`.
- If using newer than `3.3.x` version of maven, you also have to use
sequential build, Like `mvn clean install -Prelease xxxx`.
- If using lower version maven before `3.3.x`, you can use parallel
build, e.g. `mvn clean install -T 4 -Prelease xxxx`
- Not Releasing Alluxio
If you are not releasing Alluxio, you can build Alluxio like before,
please don't add the `-Prelease` into you build command line. e.g. `mvn
clean install -T 4 xxxx`

This patch adds a -Prelease. If present, it will set
createDependencyReducedPom true. The consequences are:

If you are releasing Alluxio with this profile, you are fine as long as
before.
If you are releasing Alluxio without this profile, the
`alluxio-shaded-xxx-yyy.pom` will be the same as the pom.xml under the
module directory.
If you are not releasing Alluxio but you are using this profile, you may
run into #12338
If you are not releasing Alluxio and you did not include this profile,
you are fine.
This is all documented in pom.xml and tested locally with maven 3.6.3.

pr-link: Alluxio/alluxio#12367
change-id: cid-2f26f91c9f3198d9dac11633888de0a9f8af2ce6