Fix memory leak in inode tree iterator
Currently when a user iterate through the inode tree journal entries, a
few threads are created to iterate through different directories. The
threads will be closed when the iteration completes, but if an iteration
is not fully completed, then the threads are leaked and never get
properly shutdown.

This change modifies the `JournalEntryIterable` interface to return a
closeable iterator of the journal entries, so the user of the iterator
can explicitly shutdown the threads by calling the `close()` method or
using the `try-with-resource` pattern.

The majority of the logic are in `InodeTreeBufferedIterator` class and
`CloseableIterator` class.

pr-link: Alluxio/alluxio#12383
change-id: cid-16c07efbaa0c7efea682025d3e2793415d0808c4