Prevent the grpc reader from waiting for a server response
When a grpc reader is closing, there is no message that it is expecting
from the server, however, it is currently waiting for some confirmation
from the server. This can sometimes lead to an exception in the close()
phase of the reader, which is unnecessary, since the client is
completely finished with reading.

In this change, we avoid waiting long for the server when the reader is
closing. If something does occur during the wait, we ignore it. This
avoids any unnecessary timeouts or errors during close(), after the
client is finished reading successfully.

pr-link: Alluxio/alluxio#12466
change-id: cid-33fd79578ff92a65fe9a0aa775282c3eb01c6710