[SMALLFIX] Improve block location naming
This changes `BlockStoreLocation.anyDirInTierWithMedium` to
`anyDirInAnyTierWithMedium`. The new name should be easier to
understand.

The explanation for method `anyDirInTierWithMedium` is:
```
/**
* Convenience method to return the block store location representing any
dir in any tier
* with specific medium.
*
* @param mediumType mediumType this returned block store location will
represent
* @return a BlockStoreLocation of any dir in any tier with a specific
medium
*/
```

The new name is more precise. The old name `anyDirInTierWithMedium`
makes me think it's one of the directories in a specific tier, that has
the specific medium. Rather, it means a directory in ANY tier, with the
specific medium.

pr-link: Alluxio/alluxio#12509
change-id: cid-9d69248cbab45b0953559932b25bdecab0d390f9