Use lower tiers in workers based on tier usage
This is a probabilistic approach on when we decide to assign new blocks
to lower tiers.

Before we only assign new blocks to lower tiers when tier 0 cannot hold
the initial size of a block (by default 1MB or 2MB). This leads to the
issue that most blocks are assigned to tier 0 and expand there, refusing
to go to lower tiers and freeing space on tier 0. Consequently we have
poor utilization in lower tiers.

In this change, we define heuristics that we reject allocation to
certain storage dirs if they are close to full. The closer to full it
is, the lower the chance that a block is allocated to it (and grows
there subsequently). If an allocation is rejected, we move on to trying
lower tiers and other storage dirs until we find a place.

pr-link: Alluxio/alluxio#12486
change-id: cid-2d9507a401626895cc84117b2e77b4dc59850dc7