[DOCFIX] Add flink doc back
The Flink documentation was removed a while ago

EN version was lost in
https://github.com/Alluxio/alluxio/commit/7c59f07833951131fae9f5a448753c9331714b8f
It was not moved to a new address, presumably lost in the doc structure
refactor. The history and intention of that commit has lost track.

CN version was removed in
https://github.com/Alluxio/alluxio/commit/9cb40db8171c02399fb656525b9caf13eb3f62fc#diff-4f1b02166fb51e7e45ddba897b5926d53bdc5cda08d0a76629a04e1075fdadd5
In order to keep in sync with the EN doc.

This change intends to bring the Flink doc back. Please note that this
change does not attempt to keep the Flink doc up-to-date with the
current Alluxio. That will be done in a separate PR.

pr-link: Alluxio/alluxio#12494
change-id: cid-8e07b90f0de4d8e2755f0dddc49a688f01cc61ab