Support using IP as connect host
Fix #12420 When hostname & bind.host not specified, local hostname is
chosen as the connect host of the service by default, which may cause an
UnknownHostException when the **LOCAL** service hostname cannot be
resolved by a **REMOTE** client. This problem can be avoided by
specifying alluxio.<service_name>.hostname or
alluxio.<service_name>.bind.host using an IP address. But it is really
troublesome when deploying lots of nodes with different IP addresses.

Add alluxio.service.use.ip to support using ip as the connect host when
enable.

pr-link: Alluxio/alluxio#12510
change-id: cid-8597d454940c1054785daa9118a74a206c000c87