Fix creating UFS file without a parent directory
We have been using default `CreateOptions` when creating a UFS file.
This can fail in some configuration when a file is written to an Alluxio
directory not backed by a UFS one. This change adds the option when UFS
file writer creates the UFS file so the directory is created
automatically.

https://github.com/Alluxio/alluxio/issues/12621

pr-link: Alluxio/alluxio#12618
change-id: cid-1198ff98575f126dc520a13041e3f9eabf122523