Add an option for active job count
Commands like distributedLoad and distributedCp submits jobs to the
cluster job service. The command creates 1 job for each file. By default
the command allows up to 3000 jobs to be submitted and waiting for
completion. When there are many distributedLoad/Cp commands running by
many users, there may be too many jobs in the queue for the job service.

This change intends to add a way to throttle the job submission speed at
the command side. Especially when the distributedLoad/Cp jobs are
submitted with automated scripts, we may want less than 3000 jobs to run
with each command. Maybe find a balance between speed and resource
claim, like 1000 or 500.

Volume control on Alluxio job service is beyond the scope of this PR.

Example:
```
$ bin/alluxio fs distributedLoad --job-count 1000 /
Allow 1000 active jobs
...
```

pr-link: Alluxio/alluxio#12795
change-id: cid-e0cda253f356f827a081761e48cec4393ce739dc