Support running Alluxio in intellij by import run configuration
**Is your feature request related to a problem? Please describe.**

For newbies of Alluxio, it will help them to start alluxio in the
intellij easily by import the run configuration.

Leverage this, it can make it possible to debug alluxio in the intellij
also.

**Describe the solution you'd like**

Write a script which can import the alluxio run configuration to the
alluxio project .idea folder.

**Describe alternatives you've considered**

Put the script and run configurations into dev/intellij folder

**Urgency**

Middle

**Additional context**

Everybody will like this.

pr-link: Alluxio/alluxio#12737
change-id: cid-0f4a40967cfcd59a8d47879f88746295a3f270fd