Make libjnifuse packaged in distribution
This PR packages libjnifuse (both .so and . dylib) into `lib/` dir, and
add this dir to library path of fuse process by default.
benefits:
- Users can use libjnifuse without setting java library. When binary
tarball is downloaded, JNIFuse is configured and ready to use. There
will be no difference for Mac and Linux users.
- No need to separately copy libjnifuse.so in Dockerfile.fuse --- it
will be handled as part of Alluxio tarball

@maobaolong @cheyang , FYI

pr-link: Alluxio/alluxio#12857
change-id: cid-13a9c990e850b7d60984131fa80a5535a0849bd4