Create host mount in Docker FUSE documentation
The command to ensure the bind-mounted directory on the host is created with the right permissions was still absent from the documentation. Otherwise following the tutorial could give a `Permission denied` error when trying to create the FUSE mount in the bind-mounted volume.

See also this comment: https://github.com/Alluxio/alluxio/issues/12943#issuecomment-786817860