Fix regression in AbsentUfsPathCache
Add a time stamp to indicate last refreshed time on the absent cache
entries.
Use absent cache to avoid UFS accesses when possible in the metadata
sync and load Metadata processes.

Added more comprehensive integration tests that ensures the number of
accesses to the UFS through SleepingUnderFileSystem and unit tests for
the absent cache.

fixes #12833 because now instead of checking /a, /a/b, /a/b/c for non
existent directories, we will only access the UFS once for /a because of
absent cache. We know /a/b, /a/b/c can not exist on the UFS.

Co-authored-by: David Zhu <david@alluxio.com>

pr-link: Alluxio/alluxio#12956
change-id: cid-1c836c95416214ca0256a0d62feafe78d7cf8489