Add replay statistics to Raft Journal
This change adds some new logging messages to the master while
replaying the  journal. Specifically, it will now log the rate of
journal entries applied to the state machine during the startup process.
The average rate and amortized rate are logged along with the time
elapsed since the last log message.

Additionally, there is a part of the message which attempts to estimate
the remaining entries left as well as the remaining time. This part of
the message cannot use sequence numbers to estimate the rate because
with raft, the last sequence number is non-trivial to retrieve before
starting the replay. Instead, we use the Ratis "Commit Index" which
represents batches of journal entries applied to the raft log. We are
able to get the last commit index as well as track the most recently
applied index which gives us the ability to provide the estimated time
remaining.

There is a case where an error occurs when retrieving the final commit
index. In this case, the optional variable with the index will not be
populated and the estimated time remaining statistics will not be
calculated.