Add replay statistics to the UFS journal
This commit adds new logs messages which logs statistics about each UFS
journal replay during startup. The logging happens on a periodic basis
governed by an ExponentialBackoffRetry.

A new function for the UfsJournalReader needed to be added in order to
find the highest sequence number is within a particular journal. This
allows us to get the estimated entries remaining and time left during
the replay. The method should take O(N) with the number of journal
log files. It does not read data from the log files. In the case where
there is an error determining the highest SN, an empty optional is
returned in which case the information about expected finish time and
number of entries remaining is omitted.