Stop master process upon demotion
- The code changes here are relatively simple. When the
FaultTolerantAlluxioMasterProcess changes state from primary to
secondary, then instead of switching states and continuing operation, we
should trigger the JVM to exit instead.

The reasoning behind this change is that the implementation to clean up
between state changes is not sound and leaves additional resources
hanging around which accumulate if enough state changes occur.
Defaulting to this behavior places the restart responsibility elsewhere,
but guarantees a clean slate and no resource leaks upon startup.

This is implemented with little change by simply placing a call to
stop() once the primary selector changes to the SECONDARY state
within the FTAMP. A simple new test is added to verify this behavior

Additional considerations may want to be taken into account now that
this is the default behavior. We may want to consider additional
recommended guidelines for process supervision on bare metal machines
using systemd. K8s deployments should not be affected as the default
behavior is usually to restart the pod anyway.