Add replay statistics for journal startup
This change adds some new logging messages to the master while
replaying the journal. Specifically, it will now log the rate of
journal entries applied to the state machine during the startup process.
The number of entries in a particular interval are logged along with the
current commit index and the expected entries remaining and estimated
time left.

In the embedded journal Instead, we use the Ratis "Commit Index"
which represents batches of journal entries applied to the raft log. We
are
able to get the last commit index as well as track the most recently
applied index which gives us the ability to provide the estimated time
remaining. The last SN is more readily available in the UFSJournal
since we control that implementation.

A new function for the UfsJournalReader needed to be added in order to
find the highest sequence number is within a particular journal. This
allows us to get the estimated entries remaining and time left during
the replay. The method should take O(N) with the number of journal
log files. It does not read data from the log files. In the case where
there is an error determining the highest SN, an empty optional is
returned in which case the information about expected finish time and
number of entries remaining is omitted.

There is a case where an error occurs when retrieving the final commit
index. In this case, the optional variable with the index will not be
populated and the estimated time remaining statistics will not be
calculated, only the measured rate of entries being applied.

pr-link: Alluxio/alluxio#13107
change-id: cid-1c072dd3380c5316fbe63a3fcd98c2a5c11c8416