Remove extra HashSet creation for pinned list
In `InodeTreePersistentState` we return a `Collections.unmodifiableSet`
wrapped around the pinned list. Then in `InodeTree` we create a HashSet
copy from it yet again. This is redundant.

Eventually the `getPinnedFileIds` call should be changed to a list-watch
mechanism where the worker lists all the pinned IDs then just
incrementally watch the changes. But that is not trivial and is not done
in this PR.

pr-link: Alluxio/alluxio#13228
change-id: cid-82b212f7147ac819eb78015c27cb99eed1440d5f