Fix Fuse write then read problem in async release and support umount
Previous commit
https://github.com/Alluxio/alluxio/commit/8f618358891ca89f753b369316d58221e8fb9b74
enlarges the race condition window in a access pattern of `write()` -
`release()` (async unfortunately) - `getAttr()`.

In `getAttr()`, a special logic is added to wait for the writing file to
completed before getting file size information.
`getAttr()` can happen in any cases. When it's in `write()` -
`getAttr()` - `write()` process, we don't want it to block waiting file
completed. When it's written by other threads or it's in `write()` -
`release()` - `getAttr()`, we do want it to block waiting file
completed.

This PR introducing new concurrent hash maps for the releasing entries
to fulfill the `getAttr()` logics and `umount` logics().

pr-link: Alluxio/alluxio#13160
change-id: cid-f5a6028f30287334f265b2ade21b7837e4638f6c