Change throw FuseException to log error in jni-fuse umount
If the Fuse device is busy reading/writing files, we want umount to wait
for those operations to happen.
If timeout exceeds, we want umount to be an no-op.
The current implementation is incorrect and doesn't achieve our target.
Add a TODO and change to log error for a workaround.

pr-link: Alluxio/alluxio#13231
change-id: cid-1809187450403746f15bb76fa75f349d6cd7d819