Include custom jars to lib by flag when generating tarballs
`./dev/scripts/generate-tarballs` is really convinient for alluxio
developers to generate custom alluxio tarballs. But as a custom version,
not all underfs and tool jars under ./lib are necessary to be added to
the tarballs. Maybe a custom `libjars` flag is helpful to slim the
tarball by size when generating custom tarballs.

pr-link: Alluxio/alluxio#13292
change-id: cid-05e32b43f576e9198ac16b1eb893672c21bacf88